+++ api/api-data-sharing-reporting-v1/src/main/java/com/thoughtworks/go/apiv1/datasharing/reporting/UsageStatisticsReportingControllerV1Delegate.java
@@ -22,7 +22,7 @@ import com.thoughtworks.go.api.ApiVersion;
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper;
 import com.thoughtworks.go.apiv1.datasharing.reporting.representers.UsageStatisticsReportingRepresenter;
 import com.thoughtworks.go.domain.UsageStatisticsReporting;
-import com.thoughtworks.go.server.service.DataSharingUsageStatisticsReportingService;
+import com.thoughtworks.go.server.service.datasharing.DataSharingUsageStatisticsReportingService;
 import com.thoughtworks.go.server.service.result.HttpLocalizedOperationResult;
 import com.thoughtworks.go.spark.Routes.DataSharing;
 import spark.Request;
