+++ api/api-data-sharing-settings-v1/src/test/groovy/com/thoughtworks/go/apiv1/datasharing/settings/DataSharingSettingsControllerV1DelegateTest.groovy
@@ -20,7 +20,7 @@ import com.thoughtworks.go.api.SecurityTestTrait
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper
 import com.thoughtworks.go.apiv1.datasharing.settings.representers.DataSharingSettingsRepresenter
 import com.thoughtworks.go.server.domain.DataSharingSettings
-import com.thoughtworks.go.server.service.DataSharingSettingsService
+import com.thoughtworks.go.server.service.datasharing.DataSharingSettingsService
 import com.thoughtworks.go.server.service.EntityHashingService
 import com.thoughtworks.go.spark.AdminUserSecurity
 import com.thoughtworks.go.spark.ControllerTrait
