+++ api/api-data-sharing-usage-data-v1/src/test/groovy/com/thoughtworks/go/apiv1/datasharing/usagedata/UsageStatisticsControllerV1DelegateTest.groovy
@@ -20,7 +20,7 @@ import com.thoughtworks.go.api.SecurityTestTrait
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper
 import com.thoughtworks.go.apiv1.datasharing.usagedata.representers.UsageStatisticsRepresenter
 import com.thoughtworks.go.server.domain.UsageStatistics
-import com.thoughtworks.go.server.service.DataSharingUsageDataService
+import com.thoughtworks.go.server.service.datasharing.DataSharingUsageDataService
 import com.thoughtworks.go.server.util.RSAEncryptionHelper
 import com.thoughtworks.go.spark.AdminUserSecurity
 import com.thoughtworks.go.spark.ControllerTrait
