+++ server/src/test-fast/java/com/thoughtworks/go/server/service/datasharing/DataSharingNotificationTest.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2018 ThoughtWorks, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.thoughtworks.go.server.service.datasharing;
+
+import com.thoughtworks.go.server.domain.Username;
+import com.thoughtworks.go.server.service.SecurityService;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.Mock;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.when;
+import static org.mockito.MockitoAnnotations.initMocks;
+
+public class DataSharingNotificationTest {
+    @Mock
+    private SecurityService securityService;
+    private DataSharingNotification dataSharingNotification;
+
+    @Before
+    public void setUp() throws Exception {
+        initMocks(this);
+        dataSharingNotification = new DataSharingNotification(securityService);
+    }
+
+    @Test
+    public void shouldAllowDataSharingNotificationForAdmins() {
+        Username username = new Username("foo");
+        when(securityService.isUserAdmin(username)).thenReturn(true);
+
+        assertTrue(dataSharingNotification.allowNotificationFor(username));
+    }
+
+    @Test
+    public void shouldDisAllowDataSharingNotificationForNonAdmins() throws Exception {
+        Username username = new Username("foo");
+        when(securityService.isUserAdmin(username)).thenReturn(false);
+
+        assertFalse(dataSharingNotification.allowNotificationFor(username));
+    }
+
+}
